requirements.txt
Pillow
geopy
streamlit
openai>=1.0.0
Path
datetime